dd lyway database migration support and project restructuring
- dd lyway dependencies and configuration for database migrations
- reate initial migration 1__nitial_schema.sql from existing schema
- dd sample migration 2__dd_sample_migration.sql as template
- onfigure lyway in application.yml with proper settings
- dd lywayonfig.java for pring integration
- reate .gitignore to exclude target/ and other build artifacts
- dd comprehensive project documentation (_.md)
- estructure domain classes into organized packages (lob/, policy/, productersion/)
- dd new services: alculatorervice, oefficientervice, olicyervice
- dd exception handling classes
- dd  watermark utilities
- pdate application configuration for better development experience

+++ src/main/java/ru/pt/repository/ProductVersionRepository.java
@@ -8,15 +8,15 @@ import ru.pt.domain.ProductVersion;
 
 import java.util.Optional;
 
-public interface ProductVersionRepository extends JpaRepository<ProductVersion, Long> {
+public interface ProductVersionRepository extends JpaRepository<ProductVersion, Integer> {
 
     @Query("select pv from ProductVersion pv where pv.productId = :productId and pv.versionNo = :versionNo")
-    Optional<ProductVersion> findByProductIdAndVersionNo(@Param("productId") Long productId,
+    Optional<ProductVersion> findByProductIdAndVersionNo(@Param("productId") Integer productId,
                                                          @Param("versionNo") Integer versionNo);
 
     @Modifying
     @Query("delete from ProductVersion pv where pv.productId = :productId and pv.versionNo = :versionNo")
-    int deleteByProductIdAndVersionNo(@Param("productId") Long productId,
+    int deleteByProductIdAndVersionNo(@Param("productId") Integer productId,
                                       @Param("versionNo") Integer versionNo);
 }
